requirements.txt
Flask
python-nmap
dnspython3
